Updated ome.scifio.util.FormatTools
The FormatTools class has been updated with the most recent changes and
methods.

It also is now a container for legacy FormatTools code that operates on
IFormatReaders and IFormatWriters. This unfortunate duality is
necessary during the transition period from legacy to SCIFIO formats,
but can eventually be removed.

The FormatException class in loci.formats was also updated to act as a
delegator with the ome.scifio FormatException,
to facilitate this legacy wrapping.