Updated ome.scifio.util.MetadataTools
The MetadataTools class in ome.scifio.util can now function for both the
SCIFIO components and the legacy loci.formats classes.

However, it does still need scifio equivalents for several of its
classes - but it is not clear where these methods should be, given the
impending migration of MetadataStore/Retrieve/etc... it may be more
appropriate to put them in the ome-xml component, which legacy
loci.formats code can then use for delegation.

Unfortunately, such changes can not be implemented with the current
package structure.