OmeisImporter operates on loci.formats classes
Fixed a compile error in OmeisImporter regarding the convolution between
legacy delegator classes and refactored live classes.

OmeisImporter will operate exclusively on loci.formats.IFormatReader
objects for the time being. This is due to its dependency on the Reader
wrapper extensions, such as ChannelSeparater, ChannelFiller and
fileStichter - which do not have SCIFIO equivalents.

Once those classes are converted, OmeisImporter should be refactored out
of loci.formats imports and the delegators in those classes should use
the ReaderAdapter class to delegate.