Removed protected delegation from Abstract layer
AbstractNIOHandle and AbstractService no longer delegate their protected
method functionality.

This is a tradeoff of code duplication and backwards compatibility. If
there was a downstream extension of loci.common AbstractNIOHandle, for
example, that did not have a corresponding ome.scifio version, it would
not be able to instantiate the ome.scifio AbstractNIOHandle for
delegation and thus the protected methods would be inaccessible.