ome.scifio.Metadata tracks MetadataOptions
The ome.scifio.Metadata classes will now remember what MetadataOptions
were used to generate them.

The getMetadataOptions method was added to the Metadata interface, with
an implementation in the AbstractMetadata layer.

To facilitate this tracking, Parsers will now set the MetadataOptions of
their parsed Metadata - handled automatically in the AbstractParser
layer.