Scifio translators preserve implicit Metadata
The small amount of state that is tracked by Metadata (regarding how it
was created - i.e., filtered, metadataoptions, and source) are now
maintained through translation, as these properties are common to all
Metadata implementations.

Added protected setters in the AbstractMetadata layer to facilitate the
preservation of the data, and the AbstractTranslator now automatically
carries out the transfer.

These changes also required that all AbstractTranslator extenders make a
super.translate call.