Added static AdapterTools utility
All adapters maintain a mapping between legacy and modern classes. To
avoid duplication of these classes, and thus the maps (which would not
necessarily be synchronized) it makes sense to have a single active
instance of each adapter. These instances can then be accessed via a
single static entry point.

AdapterTools serves as that entry point. It uses a static initializer to
populate its mapping of adapter classes to adapter instances, and
getAdapter(Clas<Adapter>) to query into that mapping.

Note that future adapters must be manually added to the static
initializer, or adapters need to be converted to some mechanism of
discovery.