Removed AbstractParser.metaClass
The AbstractParser class was keeping track of its Metadata class, to
avoid type erasure.

However, this information was not being used anywhere, and was private.
It is also unnecessary given that getFormat() can be used to find or
create a new, appropriate Metadata object.

Thus it was removed from the abstract and all implementating layers.