Moved ICS key, value parsing to Metadata class
Previously the ICS to OME-XML translator contained all the logic of
extracting information from the ICS key,value pairs - the
ICSFormat.Metadata class was simply a bag of these pairs.

However, there is nothing mandating that this logic be OME-XML specific.

This refactoring moves all the logic from the translator to
ICSFormat.Metadata via a series of getters. No values are actually
cached, but instead when a getter is invoked, the bag of key,value pairs
is interrogated as appropriate for the requested metadata value.

The ICS:OME-XML translator and ICS:CoreMetadata translator have been
appropriately updated to call these getters.