Added basic CoreMetadata modern->legacy support
The CoreMetadataAdapter originally did not support Modern->Legacy
conversion because of the legacy use of an array ( [] ) of CoreMetadata
objects. Without accessor methods to override there is no way to
dynamically generate information, and the only way to convert Modern
data to Legacy is pure conversion. Thus, if after conversion the Modern
CoreMetadata object is changed, the Legacy version will not reflect
those changes.

We have relaxed our assumptions to assume if a CoreMetadata object is
modified, it should be considered corrupt. Thus there is no concern in
maintaining synchronicity between the layers. As such, a modern->legacy
implementation is provided.

Note that this is a complete overhaul of the CoreMetadata adapters.
There is now:
a loci.formats.CoreMetadata[] to ome.scifio.CoreMetadata adapter
a loci.formats.CoreMetadata to ome.scifio.CoreImageMetadata adapter

and appropriate wrappers for each (the only explicit wrapper not needed
is one that isa CoreMetadata[] since it can not be extended)